refactor(playerbot): Add ResourceEventBus event extraction (5/13 complete)
Extract ResourceEvent definitions to ResourceEvents.{h,cpp}.

**Resource Migration:**
- Extracted ResourceEvent struct, ResourceEventType, ResourceEventPriority, Powers enums
- Moved 3 helper constructors (PowerChanged, PowerRegen, PowerDrained)
- Removed 76 duplicate lines from ResourceEventBus.cpp
- Updated CMakeLists.txt with new source files

**Progress:** 5/13 EventBuses migrated
**Remaining:** Aura, Auction, BotSpawn, Group, Instance, NPC, Profession, Social (8 more)
