+++ src/modules/Playerbot/Resource/ResourceEventBus.cpp
@@ -17,85 +17,6 @@
 namespace Playerbot
 {
 
-// ============================================================================
-// ResourceEvent Helper Constructors
-// ============================================================================
-
-ResourceEvent ResourceEvent::PowerChanged(ObjectGuid player, Powers type, int32 amt, int32 max)
-{
-    ResourceEvent event;
-    event.type = ResourceEventType::POWER_UPDATE;
-    event.priority = ResourceEventPriority::MEDIUM;
-    event.playerGuid = player;
-    event.powerType = type;
-    event.amount = amt;
-    event.maxAmount = max;
-    event.isRegen = false;
-    event.timestamp = std::chrono::steady_clock::now();
-    event.expiryTime = event.timestamp + std::chrono::milliseconds(5000);
-    return event;
-}
-
-ResourceEvent ResourceEvent::PowerRegen(ObjectGuid player, Powers type, int32 amt)
-{
-    ResourceEvent event;
-    event.type = ResourceEventType::POWER_UPDATE;
-    event.priority = ResourceEventPriority::LOW;
-    event.playerGuid = player;
-    event.powerType = type;
-    event.amount = amt;
-    event.maxAmount = 0;
-    event.isRegen = true;
-    event.timestamp = std::chrono::steady_clock::now();
-    event.expiryTime = event.timestamp + std::chrono::milliseconds(2000);
-    return event;
-}
-
-ResourceEvent ResourceEvent::PowerDrained(ObjectGuid player, Powers type, int32 amt)
-{
-    ResourceEvent event;
-    event.type = ResourceEventType::POWER_UPDATE;
-    event.priority = ResourceEventPriority::HIGH;
-    event.playerGuid = player;
-    event.powerType = type;
-    event.amount = -amt;
-    event.maxAmount = 0;
-    event.isRegen = false;
-    event.timestamp = std::chrono::steady_clock::now();
-    event.expiryTime = event.timestamp + std::chrono::milliseconds(10000);
-    return event;
-}
-
-bool ResourceEvent::IsValid() const
-{
-    if (type >= ResourceEventType::MAX_RESOURCE_EVENT)
-        return false;
-    if (timestamp.time_since_epoch().count() == 0)
-        return false;
-    if (playerGuid.IsEmpty())
-        return false;
-    return true;
-}
-
-bool ResourceEvent::IsExpired() const
-{
-    return std::chrono::steady_clock::now() >= expiryTime;
-}
-
-std::string ResourceEvent::ToString() const
-{
-    std::ostringstream oss;
-    oss << "[ResourceEvent] Type: " << static_cast<uint32>(type)
-        << ", Player: " << playerGuid.ToString()
-        << ", Power: " << static_cast<uint32>(powerType)
-        << ", Amount: " << amount;
-    return oss.str();
-}
-
-// ============================================================================
-// ResourceEventBus Implementation
-// ============================================================================
-
 ResourceEventBus::ResourceEventBus()
 {
     _stats.startTime = std::chrono::steady_clock::now();
