+++ src/modules/Playerbot/Resource/ResourceEventBus.h
@@ -12,10 +12,8 @@
 
 #include "Define.h"
 #include "Threading/LockHierarchy.h"
-#include "ObjectGuid.h"
+#include "ResourceEvents.h"
 #include "Core/DI/Interfaces/IResourceEventBus.h"
-#include <chrono>
-#include <string>
 #include <vector>
 #include <unordered_map>
 #include <queue>
@@ -26,60 +24,6 @@ namespace Playerbot
 {
 
 class BotAI;
-
-enum class ResourceEventType : uint8
-{
-    HEALTH_UPDATE = 0,
-    POWER_UPDATE,
-    BREAK_TARGET,
-    MAX_RESOURCE_EVENT
-};
-
-enum class ResourceEventPriority : uint8
-{
-    CRITICAL = 0,
-    HIGH = 1,
-    MEDIUM = 2,
-    LOW = 3,
-    BATCH = 4
-};
-
-enum class Powers : uint8
-{
-    POWER_MANA = 0,
-    POWER_RAGE = 1,
-    POWER_FOCUS = 2,
-    POWER_ENERGY = 3,
-    POWER_RUNIC_POWER = 6
-};
-
-struct ResourceEvent
-{
-    ResourceEventType type;
-    ResourceEventPriority priority;
-    ObjectGuid playerGuid;
-    Powers powerType;
-    int32 amount;
-    int32 maxAmount;
-    bool isRegen;
-    std::chrono::steady_clock::time_point timestamp;
-    std::chrono::steady_clock::time_point expiryTime;
-
-    bool IsValid() const;
-    bool IsExpired() const;
-    std::string ToString() const;
-
-    // Helper constructors
-    static ResourceEvent PowerChanged(ObjectGuid player, Powers type, int32 amt, int32 max);
-    static ResourceEvent PowerRegen(ObjectGuid player, Powers type, int32 amt);
-    static ResourceEvent PowerDrained(ObjectGuid player, Powers type, int32 amt);
-
-    // Priority comparison for priority queue
-    bool operator<(ResourceEvent const& other) const
-    {
-        return priority > other.priority;
-    }
-};
 
 class TC_GAME_API ResourceEventBus final : public IResourceEventBus
 {
